+++ server/cover.out
@@ -1,82 +1,1 @@
 mode: set
-server/user/repository/memory.go:13.48,22.16 7 0
-server/user/repository/memory.go:26.2,28.3 1 0
-server/user/repository/memory.go:22.16,24.3 1 0
-server/user/repository/memory.go:37.63,40.16 3 1
-server/user/repository/memory.go:43.2,44.18 2 1
-server/user/repository/memory.go:52.2,52.19 1 1
-server/user/repository/memory.go:40.16,42.3 1 1
-server/user/repository/memory.go:44.18,47.17 3 1
-server/user/repository/memory.go:50.3,50.32 1 1
-server/user/repository/memory.go:47.17,49.4 1 1
-server/user/repository/memory.go:56.72,61.16 3 1
-server/user/repository/memory.go:64.2,64.20 1 1
-server/user/repository/memory.go:61.16,63.3 1 1
-server/user/repository/memory.go:68.86,73.16 3 1
-server/user/repository/memory.go:76.2,76.20 1 1
-server/user/repository/memory.go:73.16,75.3 1 1
-server/user/repository/memory.go:80.70,89.16 3 1
-server/user/repository/memory.go:92.2,92.30 1 0
-server/user/repository/memory.go:89.16,91.3 1 1
-server/user/repository/memory.go:96.70,110.16 3 1
-server/user/repository/memory.go:113.2,113.30 1 0
-server/user/repository/memory.go:110.16,112.3 1 1
-server/user/repository/memory.go:117.62,122.16 2 1
-server/user/repository/memory.go:125.2,125.30 1 1
-server/user/repository/memory.go:122.16,124.3 1 1
-server/user/delivery/http/handlers.go:23.52,35.2 9 0
-server/user/delivery/http/handlers.go:37.57,40.47 2 1
-server/user/delivery/http/handlers.go:44.2,50.16 5 1
-server/user/delivery/http/handlers.go:55.2,57.59 2 1
-server/user/delivery/http/handlers.go:61.2,63.12 2 1
-server/user/delivery/http/handlers.go:40.47,42.3 1 0
-server/user/delivery/http/handlers.go:50.16,53.3 2 0
-server/user/delivery/http/handlers.go:57.59,59.3 1 0
-server/user/delivery/http/handlers.go:66.57,70.50 2 1
-server/user/delivery/http/handlers.go:74.2,79.16 4 1
-server/user/delivery/http/handlers.go:83.2,85.75 2 1
-server/user/delivery/http/handlers.go:89.2,92.57 2 1
-server/user/delivery/http/handlers.go:96.2,96.12 1 1
-server/user/delivery/http/handlers.go:70.50,72.3 1 0
-server/user/delivery/http/handlers.go:79.16,81.3 1 0
-server/user/delivery/http/handlers.go:85.75,87.3 1 0
-server/user/delivery/http/handlers.go:92.57,94.3 1 0
-server/user/delivery/http/handlers.go:99.60,105.26 4 1
-server/user/delivery/http/handlers.go:114.2,114.12 1 1
-server/user/delivery/http/handlers.go:105.26,112.3 3 1
-server/user/delivery/http/handlers.go:117.53,119.2 1 0
-server/user/delivery/http/handlers.go:121.64,125.59 2 1
-server/user/delivery/http/handlers.go:129.2,133.16 3 1
-server/user/delivery/http/handlers.go:137.2,139.45 3 1
-server/user/delivery/http/handlers.go:143.2,143.45 1 1
-server/user/delivery/http/handlers.go:147.2,148.16 2 1
-server/user/delivery/http/handlers.go:151.2,153.71 2 1
-server/user/delivery/http/handlers.go:157.2,157.12 1 1
-server/user/delivery/http/handlers.go:125.59,127.3 1 0
-server/user/delivery/http/handlers.go:133.16,135.3 1 0
-server/user/delivery/http/handlers.go:139.45,141.3 1 0
-server/user/delivery/http/handlers.go:143.45,145.3 1 0
-server/user/delivery/http/handlers.go:148.16,150.3 1 0
-server/user/delivery/http/handlers.go:153.71,155.3 1 0
-server/user/delivery/http/handlers.go:160.62,165.16 3 0
-server/user/delivery/http/handlers.go:169.2,174.16 4 0
-server/user/delivery/http/handlers.go:178.2,183.16 5 0
-server/user/delivery/http/handlers.go:186.2,188.57 2 0
-server/user/delivery/http/handlers.go:192.2,192.12 1 0
-server/user/delivery/http/handlers.go:165.16,167.3 1 0
-server/user/delivery/http/handlers.go:174.16,176.3 1 0
-server/user/delivery/http/handlers.go:183.16,185.3 1 0
-server/user/delivery/http/handlers.go:188.57,190.3 1 0
-server/user/delivery/http/handlers.go:195.68,197.16 2 0
-server/user/delivery/http/handlers.go:202.2,203.16 2 0
-server/user/delivery/http/handlers.go:207.2,210.16 3 0
-server/user/delivery/http/handlers.go:214.2,216.44 2 0
-server/user/delivery/http/handlers.go:221.2,221.27 1 0
-server/user/delivery/http/handlers.go:197.16,200.3 2 0
-server/user/delivery/http/handlers.go:203.16,206.3 2 0
-server/user/delivery/http/handlers.go:210.16,213.3 2 0
-server/user/delivery/http/handlers.go:216.44,219.3 2 0
-server/user/delivery/http/handlers.go:224.61,231.32 2 1
-server/user/delivery/http/handlers.go:235.2,235.45 1 1
-server/user/delivery/http/handlers.go:231.32,233.3 1 0
-server/user/delivery/http/handlers.go:238.57,241.2 2 1
